OAcc:qAdd kata containers support for starlingx with kata-community released runtime binaries, static linked kernel and images.
As in previous StarlingX 7.0 Debian and CentOS Release we have kata container
support which is from community kata 1.x release.

To leverage the OS community effort with the latest kata container version, need
to utilize the prebuild runtime binaries with static linked kernel and images
for the Guest VM, by default the hypervisor is amd64-x86 qemu.

To provide the customization support for StarlingX, the kata container default
configuration file will be modified.

Story: 2010765
Task: 48073

Change-Id: I7b8a0cab1e71e65291792b763ca801480648b511
Signed-off-by: David Liu <david.liu@windriver.com>